requirements.txt
mlflow==2.14.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
flask==3.0.3
gunicorn==22.0.0
pyyaml==6.0.2
dvc==3.54.1
dvc[gdrive]==3.54.1
requests==2.32.3
